+++ AlphaToe/game_eval.py
@@ -8,12 +8,14 @@ from dotenv import load_dotenv
 import os
 import pandas as pd
 from pprint import pprint
+import helpers
 
 
 load_dotenv()
 xml_file = f"{os.getenv('ROOT_PATH')}/{os.getenv('PORTSCAN_XML')}"
 commands_csv = f"{os.getenv('ROOT_PATH')}/{os.getenv('COMMANDS_CSV')}"
 attacks = pd.read_csv(commands_csv)
+
 
 metasploit = {
     0: "scan_command",
@@ -39,10 +41,13 @@ def show_intention(attack_id, attacks=pd.read_csv(commands_csv)):
 # **returns**: List[int].
 
 def get_state_mapping_evaluation(prev_state: List[List[int]], current_state: List[List[int]], debug: bool = False):
-    exploit_file_3x3 = "exploit_3x3"
-    exploit_file_5x5 = "exploit_5x5"
-    set_file_5x5 = "set_5x5"
-    mp.write_logging_files(exploit_file_3x3, exploit_file_5x5, set_file_5x5)
+    # exploit_file_3x3 = "exploit_3x3"
+    # exploit_file_5x5 = "exploit_5x5"
+    # set_file_5x5 = "set_5x5"
+    exploit_file_3x3 = os.getenv('EXPLOIT_3x3')
+    exploit_file_5x5 = os.getenv('EXPLOIT_5x5')
+    set_file_5x5 = os.getenv('SET_5x5')
+    helpers.write_logging_files(exploit_file_3x3, exploit_file_5x5, set_file_5x5)
     move = mp.get_latest_move(prev_state, current_state)
     i = move[0]
     j = move[1]
